[FIX] account_reports: prevent reset to draft if subsequent closing entries
Prevents the user to reset to draft a tax closing entry if subsequent closing entries are already posted. This way, the user is forced to reset to draft progressively back in time the closing entries. Accounting wise, this case should not happen. It is by the way prevented by the irreversible_lock_date module (which can be however uninstalled). This mechanism solves an issue with carryovers that need to be progressively recomputed when resetting to draft a closing entry.

task-3520338

closes odoo/enterprise#59534

X-original-commit: a5afa4b21d47b1b82a4a4ce66d67d3cd6e6585f1

+++ account_reports/models/account_move.py
@@ -65,7 +65,21 @@ class AccountMove(models.Model):
                 raise UserError(_("You cannot reset this closing entry to draft, as it would delete carryover values impacting the tax report of a "
                                   "locked period. To do this, you first need to modify you tax return lock date."))
 
+            if self._has_subsequent_posted_closing_moves():
+                raise UserError(_("You cannot reset this closing entry to draft, as another closing entry has been posted at a later date."))
+
             carryover_values.unlink()
+
+    def _has_subsequent_posted_closing_moves(self):
+        self.ensure_one()
+        closing_domains = [
+            ('company_id', '=', self.company_id.id),
+            ('tax_closing_end_date', '!=', False),
+            ('state', '=', 'posted'),
+            ('date', '>', self.date),
+            ('fiscal_position_id', '=', self.fiscal_position_id.id)
+        ]
+        return bool(self.env['account.move'].search_count(closing_domains, limit=1))
 
     def action_open_tax_report(self):
         action = self.env["ir.actions.actions"]._for_xml_id("account_reports.action_account_report_gt")
